# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NCST LMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses/{courseId}/modules:
    get:
      operationId: listModules
      tags: [modules]
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Modules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Module"
    post:
      operationId: createModule
      tags: [modules]
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModuleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"
  /modules/{id}:
    patch:
      operationId: updateModule
      tags: [modules]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModuleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"
    delete:
      operationId: deleteModule
      tags: [modules]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── FILES / CONTENT ────────────────────────────────────────────────────────
  /modules/{moduleId}/files:
    get:
      operationId: listFiles
      tags: [files]
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CourseFile"
  /files:
    post:
      operationId: createFile
      tags: [files]
      summary: Upload file metadata (multipart handled separately)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseFile"
  /files/{id}:
    delete:
      operationId: deleteFile
      tags: [files]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted


  # ─── QUIZZES / EXAMS ────────────────────────────────────────────────────────
...
  /quizzes/{quizId}/questions:
    get:
      operationId: listQuestions
      tags: [questions]
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
    post:
      operationId: createQuestion
      tags: [questions]
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
  /questions/{id}:
    patch:
      operationId: updateQuestion
      tags: [questions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
    delete:
      operationId: deleteQuestion
      tags: [questions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── SUBMISSIONS ─────────────────────────────────────────────────────────────
...
  /attempts/{attemptId}/grade:
    patch:
      operationId: gradeAttempt
      tags: [grades]
      summary: Teacher manually grades attempt
      parameters:
        - name: attemptId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GradeInput"
      responses:
        "200":
          description: Graded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizAttempt"
  /courses/{courseId}/grades:
    get:
      operationId: listCourseGrades
      tags: [grades]
      summary: List all grades for a course (for Excel export)
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Grades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GradeRow"
  /courses/{courseId}/grades/export:
    get:
      operationId: exportGrades
      tags: [grades]
      summary: Export grades as data for Excel
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Export data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GradeExport"

  # ─── ALERTS / PROCTORING ─────────────────────────────────────────────────────
...
components:
  schemas:
...
    Module:
      type: object
      required: [id, courseId, title, position, createdAt]
      properties:
        id:
          type: integer
        courseId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        position:
          type: integer
        createdAt:
          type: string
          format: date-time
    ModuleInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
        position:
          type: integer
    ModuleUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        position:
          type: integer

    # ─── FILE ─────────────────────────────────────────────────────────────────
    CourseFile:
      type: object
      required: [id, moduleId, fileName, fileType, uploadedBy, createdAt]
      properties:
        id:
          type: integer
        moduleId:
          type: integer
        fileName:
          type: string
        originalName:
          type: string
        fileType:
          type: string
        fileSize:
          type: ["integer", "null"]
        url:
          type: string
        uploadedBy:
          type: integer
        uploaderName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    FileInput:
      type: object
      required: [moduleId, fileName, fileType, url, uploadedBy]
      properties:
        moduleId:
          type: integer
        fileName:
          type: string
        originalName:
          type: string
        fileType:
          type: string
        fileSize:
          type: integer
        url:
          type: string
        uploadedBy:
          type: integer

    # ─── QUIZ ─────────────────────────────────────────────────────────────────
...
    QuizAttempt:
      type: object
      required: [id, quizId, studentId, status, startedAt]
      properties:
        id:
          type: integer
        quizId:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        quizTitle:
          type: ["string", "null"]
        status:
          type: string
          enum: [in_progress, submitted, force_submitted, graded]
        score:
          type: ["number", "null"]
        maxScore:
          type: ["number", "null"]
        answers:
          type: ["string", "null"]
        feedback:
          type: ["string", "null"]
        startedAt:
          type: string
          format: date-time
        submittedAt:
          type: ["string", "null"]
          format: date-time
        lateEntry:
          type: boolean
        lateMinutes:
          type: ["integer", "null"]

    # ─── QUESTION ─────────────────────────────────────────────────────────────
    Question:
      type: object
      required: [id, quizId, questionText, questionType, points, position]
      properties:
        id:
          type: integer
        quizId:
          type: integer
        questionText:
          type: string
        questionType:
          type: string
          enum: [multiple_choice, true_false, short_answer, essay, file_upload]
        points:
          type: integer
        position:
          type: integer
        options:
          type: ["string", "null"]
        correctAnswer:
          type: ["string", "null"]
        explanation:
          type: ["string", "null"]
    QuestionInput:
      type: object
      required: [questionText, questionType, points, position]
      properties:
        questionText:
          type: string
        questionType:
          type: string
          enum: [multiple_choice, true_false, short_answer, essay, file_upload]
        points:
          type: integer
        position:
          type: integer
        options:
          type: string
        correctAnswer:
          type: string
        explanation:
          type: string
    QuestionUpdate:
      type: object
      properties:
        questionText:
          type: string
        questionType:
          type: string
          enum: [multiple_choice, true_false, short_answer, essay, file_upload]
        points:
          type: integer
        position:
          type: integer
        options:
          type: string
        correctAnswer:
          type: string
        explanation:
          type: string

    # ─── SUBMISSION ───────────────────────────────────────────────────────────
...
    GradeInput:
      type: object
      required: [score]
      properties:
        score:
          type: number
        feedback:
          type: string
    GradeRow:
      type: object
      required: [studentId, studentName, studentEmail, quizTitle, score, maxScore, status, submittedAt]
      properties:
        studentId:
          type: integer
        studentName:
          type: string
        studentEmail:
          type: string
        quizTitle:
          type: string
        score:
          type: ["number", "null"]
        maxScore:
          type: ["number", "null"]
        status:
          type: string
        submittedAt:
          type: ["string", "null"]
    GradeExport:
      type: object
      required: [courseName, rows]
      properties:
        courseName:
          type: string
        rows:
          type: array
          items:
            $ref: "#/components/schemas/GradeRow"

    # ─── ALERT ────────────────────────────────────────────────────────────────
...