# lib/api-zod/src/generated/types/questionUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NCST LMS API
 * OpenAPI spec version: 0.1.0
 */
import type { QuestionUpdateQuestionType } from "./questionUpdateQuestionType";

export interface QuestionUpdate {
  questionText?: string;
  questionType?: QuestionUpdateQuestionType;
  points?: number;
  position?: number;
  options?: string;
  correctAnswer?: string;
  explanation?: string;
}